polarity: added module for composite, use pyrocko stations as data
